Move Python plugin into packages/minicode-plugin-python workspace
Establishes the canonical layout for a minicode language plugin. The
plugin used to live in src/indexer/plugins/python.ts as part of the root
package; it now lives in its own private workspace package mirroring the
shape an external `minicode-plugin-*` author would publish.

Structure:

  packages/minicode-plugin-python/
    package.json     — name: minicode-plugin-python, deps: tree-sitter,
                       tree-sitter-python; peerDep: @minicode/agent-sdk
    tsconfig.json    — same compiler settings as agent-sdk
    src/index.ts     — the plugin (imports LanguagePlugin from
                       @minicode/agent-sdk, not from a relative path)
    tests/           — pure unit tests of pythonPlugin.indexFile
                       (the kind of tests an external author would run
                       in isolation via `npm test` inside the package)

Root changes:

- `tree-sitter` and `tree-sitter-python` move out of root dependencies
  and into the plugin package's dependencies.
- `minicode-plugin-python: "*"` added to root dependencies and
  bundleDependencies, so `npm install -g @sean.holung/minicode` still
  pulls the plugin transparently — no separate user-facing install.
- Plugin loader's dynamic import switches from `./plugins/python.js`
  to `minicode-plugin-python` (the npm package name). Same graceful
  degradation if the native build is unavailable.
- Build/pretest scripts add `npm run build --workspace=packages/
  minicode-plugin-python` to the chain.
- Root `npm test` glob expanded to include `packages/*/tests/**/*.test.ts`
  so package tests run alongside root tests in one report.
- Root `tests/python-plugin.test.ts` keeps integration tests
  (loadPlugins, getPluginForFile, buildProjectIndex, mixed TS+Python,
  .pyi indexing); pure indexFile unit tests moved into the package.

PLUGIN_SPEC.md updated to point at the new location and explain the
workspace-package shape.

Total tests: 560/562 passing (the 2 failures are the same pre-existing
sandbox `git commit` config issue, unrelated). Lint clean, typecheck
clean.
