+++ package.json
@@ -27,6 +27,7 @@
     "dotenv",
     "ink",
     "marked",
+    "minicode-plugin-python",
     "picocolors",
     "react",
     "typescript",
@@ -37,12 +38,12 @@
     "build:web": "node scripts/build-web.mjs",
     "dev": "node scripts/build-web.mjs && node --import tsx src/index.ts",
     "dev:ink": "CLI_UI_MODE=ink node --import tsx src/index.ts",
-    "build": "npm run build --workspace=packages/agent-sdk && tsc -p tsconfig.json && node scripts/build-web.mjs && chmod +x dist/src/index.js",
+    "build": "npm run build --workspace=packages/agent-sdk && npm run build --workspace=packages/minicode-plugin-python && tsc -p tsconfig.json && node scripts/build-web.mjs && chmod +x dist/src/index.js",
     "start": "node dist/src/index.js",
     "install:global": "npm run build && npm link",
     "lint": "eslint \"src/**/*.ts\" \"src/**/*.tsx\" \"tests/**/*.ts\" --max-warnings=0",
-    "pretest": "npm run build --workspace=packages/agent-sdk && node scripts/build-web.mjs",
-    "test": "node --test --import tsx \"tests/**/*.test.ts\"",
+    "pretest": "npm run build --workspace=packages/agent-sdk && npm run build --workspace=packages/minicode-plugin-python && node scripts/build-web.mjs",
+    "test": "node --test --import tsx \"tests/**/*.test.ts\" \"packages/*/tests/**/*.test.ts\"",
     "verify-index": "node --import tsx test-programs/verify-index/verify.ts",
     "benchmark": "node --import tsx scripts/run-benchmarks.ts"
   },
@@ -54,10 +55,9 @@
     "dotenv": "^17.3.1",
     "ink": "^6.8.0",
     "marked": "^17.0.5",
+    "minicode-plugin-python": "*",
     "picocolors": "^1.1.1",
     "react": "^19.2.4",
-    "tree-sitter": "^0.25.0",
-    "tree-sitter-python": "^0.25.0",
     "typescript": "^5.9.3",
     "ws": "^8.19.0",
     "zod": "^3.25.76"
