+++ packages/minicode-plugin-python/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "minicode-plugin-python",
+  "version": "0.1.0",
+  "private": true,
+  "description": "Python language plugin for minicode (built-in)",
+  "type": "module",
+  "engines": {
+    "node": ">=22.0.0"
+  },
+  "main": "dist/src/index.js",
+  "types": "dist/src/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/src/index.d.ts",
+      "import": "./dist/src/index.js"
+    }
+  },
+  "files": ["dist", "README.md", "LICENSE"],
+  "scripts": {
+    "build": "tsc -p tsconfig.json",
+    "test": "node --test --import tsx \"tests/**/*.test.ts\"",
+    "lint": "eslint \"src/**/*.ts\" \"tests/**/*.ts\" --max-warnings=0"
+  },
+  "dependencies": {
+    "tree-sitter": "^0.25.0",
+    "tree-sitter-python": "^0.25.0"
+  },
+  "peerDependencies": {
+    "@minicode/agent-sdk": "*"
+  },
+  "devDependencies": {
+    "@minicode/agent-sdk": "*",
+    "@types/node": "^25.2.3",
+    "tsx": "^4.21.0",
+    "typescript": "^5.9.3"
+  }
+}
