+++ tests/python-plugin.test.ts
@@ -6,162 +6,6 @@ import { test } from "node:test";
 
 import { getPluginForFile, loadPlugins } from "../src/indexer/plugin-loader.js";
 import { buildProjectIndex } from "../src/indexer/project-index.js";
-import { pythonPlugin } from "../src/indexer/plugins/python.js";
-
-const SAMPLE_PY = [
-  "import os",
-  "from typing import Protocol",
-  "",
-  "@dataclass",
-  "class Animal:",
-  '    """An animal."""',
-  "    name: str",
-  "",
-  "    def __init__(self, name: str):",
-  "        self.name = name",
-  "",
-  "    async def fetch(self, ball: int) -> bool:",
-  "        return True",
-  "",
-  "class Dog(Animal):",
-  "    def speak(self) -> str:",
-  '        """Bark!"""',
-  '        return "woof"',
-  "",
-  "",
-  "def helper(x: int) -> str:",
-  "    return str(x)",
-  "",
-  "_private = 1",
-  "",
-  "type Vec = list[float]",
-  "",
-].join("\n");
-
-test("Python plugin extracts classes, methods, functions, type aliases", () => {
-  const symbols = pythonPlugin.indexFile("sample.py", SAMPLE_PY);
-  const names = symbols.map((s) => s.qualifiedName);
-  assert.ok(names.includes("Animal"), "should extract class Animal");
-  assert.ok(names.includes("Animal.__init__"), "should extract method __init__");
-  assert.ok(names.includes("Animal.fetch"), "should extract async method fetch");
-  assert.ok(names.includes("Dog"), "should extract class Dog");
-  assert.ok(names.includes("Dog.speak"), "should extract method Dog.speak");
-  assert.ok(names.includes("helper"), "should extract top-level helper");
-  assert.ok(names.includes("Vec"), "should extract PEP 695 type alias");
-});
-
-test("Python plugin marks async def in signature", () => {
-  const symbols = pythonPlugin.indexFile("sample.py", SAMPLE_PY);
-  const fetch = symbols.find((s) => s.qualifiedName === "Animal.fetch");
-  assert.ok(fetch);
-  assert.ok(
-    fetch!.signature.startsWith("async def fetch"),
-    `expected async-def signature, got ${fetch!.signature}`,
-  );
-});
-
-test("Python plugin extracts class-superclass header in signature", () => {
-  const symbols = pythonPlugin.indexFile("sample.py", SAMPLE_PY);
-  const dog = symbols.find((s) => s.qualifiedName === "Dog");
-  assert.ok(dog);
-  assert.equal(dog!.signature, "class Dog(Animal)");
-});
-
-test("Python plugin extracts docstrings as docComment", () => {
-  const symbols = pythonPlugin.indexFile("sample.py", SAMPLE_PY);
-  const animal = symbols.find((s) => s.qualifiedName === "Animal");
-  const speak = symbols.find((s) => s.qualifiedName === "Dog.speak");
-  assert.ok(animal);
-  assert.ok(speak);
-  assert.equal(animal!.docComment, "An animal.");
-  assert.equal(speak!.docComment, "Bark!");
-});
-
-test("Python plugin treats leading-underscore names as not exported", () => {
-  const code = [
-    "def _internal():",
-    "    pass",
-    "",
-    "def public():",
-    "    pass",
-    "",
-    "def __dunder__():",
-    "    pass",
-    "",
-  ].join("\n");
-  const symbols = pythonPlugin.indexFile("sample.py", code);
-  assert.equal(
-    symbols.find((s) => s.name === "_internal")!.exported,
-    false,
-    "underscored names are private",
-  );
-  assert.equal(symbols.find((s) => s.name === "public")!.exported, true);
-  assert.equal(
-    symbols.find((s) => s.name === "__dunder__")!.exported,
-    true,
-    "dunders are exported",
-  );
-});
-
-test("Python plugin uses outer range for decorated definitions (decorators included in startLine)", () => {
-  const code = [
-    "@dataclass",
-    "@frozen",
-    "class C:",
-    "    pass",
-    "",
-  ].join("\n");
-  const symbols = pythonPlugin.indexFile("sample.py", code);
-  const c = symbols.find((s) => s.name === "C");
-  assert.ok(c);
-  assert.equal(c!.startLine, 1, "startLine should include decorators");
-});
-
-test("Python plugin handles nested classes with correct qualified names", () => {
-  const code = [
-    "class Outer:",
-    "    class Inner:",
-    "        def m(self):",
-    "            pass",
-    "    def o(self):",
-    "        pass",
-    "",
-  ].join("\n");
-  const symbols = pythonPlugin.indexFile("nested.py", code);
-  const names = symbols.map((s) => s.qualifiedName).sort();
-  assert.deepEqual(names, ["Outer", "Outer.Inner", "Outer.Inner.m", "Outer.o"]);
-});
-
-test("Python plugin extracts PEP 695 type alias signature", () => {
-  const code = "type Vec = list[float]\n";
-  const symbols = pythonPlugin.indexFile("sample.py", code);
-  const vec = symbols.find((s) => s.name === "Vec");
-  assert.ok(vec);
-  assert.equal(vec!.kind, "type");
-  assert.equal(vec!.signature, "type Vec = list[float]");
-});
-
-test("Python plugin returns empty array for empty file", () => {
-  const symbols = pythonPlugin.indexFile("empty.py", "");
-  assert.equal(symbols.length, 0);
-});
-
-test("Python plugin handles malformed syntax without throwing", () => {
-  const symbols = pythonPlugin.indexFile("bad.py", "def : broken\n");
-  assert.ok(Array.isArray(symbols));
-});
-
-test("Python plugin skips module docstrings", () => {
-  const code = [
-    '"""Module docstring."""',
-    "def f():",
-    "    pass",
-    "",
-  ].join("\n");
-  const symbols = pythonPlugin.indexFile("sample.py", code);
-  assert.equal(symbols.length, 1);
-  assert.equal(symbols[0]!.name, "f");
-});
 
 test("Plugin loader returns the built-in Python plugin", async () => {
   const plugins = await loadPlugins("/tmp");
@@ -334,87 +178,6 @@ test("Python plugin calls edges resolve absolute imports", async () => {
   );
 });
 
-test("Python plugin includes decorators in signatures", () => {
-  const code = [
-    "@dataclass(frozen=True)",
-    "class Foo:",
-    "    pass",
-    "",
-    "@property",
-    "def value() -> int:",
-    "    return 1",
-    "",
-  ].join("\n");
-  const symbols = pythonPlugin.indexFile("sample.py", code);
-  const foo = symbols.find((s) => s.name === "Foo");
-  const value = symbols.find((s) => s.name === "value");
-  assert.ok(foo);
-  assert.ok(value);
-  assert.ok(
-    foo!.signature.includes("@dataclass(frozen=True)"),
-    `expected @dataclass in signature, got ${foo!.signature}`,
-  );
-  assert.ok(
-    value!.signature.startsWith("@property"),
-    `expected @property in signature, got ${value!.signature}`,
-  );
-});
-
-test("Python plugin honors __all__ for top-level exported flag", () => {
-  const code = [
-    "__all__ = ['public_helper', '_private_but_listed']",
-    "",
-    "def public_helper():",
-    "    pass",
-    "",
-    "def not_listed():",
-    "    pass",
-    "",
-    "def _private_but_listed():",
-    "    pass",
-    "",
-    "def _private():",
-    "    pass",
-    "",
-  ].join("\n");
-  const symbols = pythonPlugin.indexFile("sample.py", code);
-  const get = (name: string) =>
-    symbols.find((s) => s.name === name) ?? assert.fail(`missing ${name}`);
-  assert.equal(get("public_helper").exported, true);
-  assert.equal(get("not_listed").exported, false, "not in __all__ → not exported");
-  assert.equal(
-    get("_private_but_listed").exported,
-    true,
-    "underscored but in __all__ → exported",
-  );
-  assert.equal(get("_private").exported, false);
-});
-
-test("Python plugin marks Protocol classes as interfaces", () => {
-  const code = [
-    "from typing import Protocol",
-    "",
-    "class Animal(Protocol):",
-    "    def speak(self) -> str: ...",
-    "",
-    "class GenericIface(Protocol[int]):",
-    "    pass",
-    "",
-    "import typing",
-    "class QualifiedIface(typing.Protocol):",
-    "    pass",
-    "",
-    "class Plain:",
-    "    pass",
-    "",
-  ].join("\n");
-  const symbols = pythonPlugin.indexFile("sample.py", code);
-  assert.equal(symbols.find((s) => s.name === "Animal")!.kind, "interface");
-  assert.equal(symbols.find((s) => s.name === "GenericIface")!.kind, "interface");
-  assert.equal(symbols.find((s) => s.name === "QualifiedIface")!.kind, "interface");
-  assert.equal(symbols.find((s) => s.name === "Plain")!.kind, "class");
-});
-
 test("Python plugin Protocol→interface still receives extends edges as base class", async () => {
   const workspaceRoot = await mkdtemp(path.join(tmpdir(), "minicode-py-proto-"));
   await writeFile(
@@ -494,7 +257,6 @@ test("buildProjectIndex handles a mixed TypeScript + Python workspace", async ()
 
   const index = await buildProjectIndex(workspaceRoot);
 
-  // TypeScript symbols
   const greeter = index.getSymbol("Greeter");
   const tsService = index.getSymbol("Service");
   assert.ok(greeter, "TypeScript interface Greeter should be indexed");
@@ -502,7 +264,6 @@ test("buildProjectIndex handles a mixed TypeScript + Python workspace", async ()
   assert.ok(tsService, "TypeScript class Service should be indexed");
   assert.equal(tsService!.kind, "class");
 
-  // Python symbols
   const user = index.getSymbol("User");
   const makeUser = index.getSymbol("make_user");
   assert.ok(user, "Python User dataclass should be indexed");
@@ -510,7 +271,6 @@ test("buildProjectIndex handles a mixed TypeScript + Python workspace", async ()
   assert.ok(makeUser, "Python make_user function should be indexed");
   assert.equal(makeUser!.filePath, "src/service.py");
 
-  // Edges from BOTH plugins should land in the same graph
   const tsImpl = index.dependencyEdges.find(
     (e) => e.kind === "implements" && e.from === "Service" && e.to === "Greeter",
   );
